[ESP32] move route_hook to platform layer
