+++ src/platform/ESP32/ConnectivityManagerImpl_WiFi.cpp
@@ -29,6 +29,7 @@
 #include <platform/ESP32/ESP32Utils.h>
 #include <platform/ESP32/NetworkCommissioningDriver.h>
 #include <platform/internal/BLEManager.h>
+#include <platform/ESP32/route_hook/ESP32RouteHook.h>
 
 #include "esp_event.h"
 #include "esp_netif.h"
@@ -1094,6 +1095,8 @@ void ConnectivityManagerImpl::OnIPv6AddressAvailable(const ip_event_got_ip6_t & 
     event.Type                           = DeviceEventType::kInterfaceIpAddressChanged;
     event.InterfaceIpAddressChanged.Type = InterfaceIpChangeType::kIpV6_Assigned;
     PlatformMgr().PostEventOrDie(&event);
+
+    esp_route_hook_init(esp_netif_get_handle_from_ifkey("WIFI_STA_DEF"));
 }
 
 } // namespace DeviceLayer
