+++ src/platform/ESP32/route_hook/ESP32RouteTable.h
@@ -12,6 +12,8 @@
 extern "C" {
 #endif
 
+#define TAG "ROUTE_HOOK"
+
 /**
  * @brief Route table entry
  *
